feat(apiserver): change gorm v1 to v2

+++ internal/apiserver/api/v1/cache/cache.go
@@ -24,8 +24,8 @@ type Cache struct {
 // ListSecrets returns all secrets.
 func (c Cache) ListSecrets(ctx context.Context, r *pb.ListSecretsRequest) (*pb.ListSecretsResponse, error) {
 	opts := metav1.ListOptions{
-		Offset: pointer.ToInt64(0),
-		Limit:  pointer.ToInt64(-1),
+		Offset: pointer.ToInt(0),
+		Limit:  pointer.ToInt(-1),
 	}
 
 	secrets, err := store.Client().Secrets().List("", opts)
@@ -55,8 +55,8 @@ func (c Cache) ListSecrets(ctx context.Context, r *pb.ListSecretsRequest) (*pb.L
 // ListPolicies returns all policies.
 func (c Cache) ListPolicies(ctx context.Context, r *pb.ListPoliciesRequest) (*pb.ListPoliciesResponse, error) {
 	opts := metav1.ListOptions{
-		Offset: pointer.ToInt64(0),
-		Limit:  pointer.ToInt64(-1),
+		Offset: pointer.ToInt(0),
+		Limit:  pointer.ToInt(-1),
 	}
 
 	policies, err := store.Client().Policies().List("", opts)
